Refactoring. Making sure contour related resources are removed when not in use.
