+++ controllers/kafkacluster_controller.go
@@ -41,7 +41,7 @@ import (
 	"github.com/banzaicloud/koperator/pkg/kafkaclient"
 	"github.com/banzaicloud/koperator/pkg/pki"
 	"github.com/banzaicloud/koperator/pkg/resources"
-	"github.com/banzaicloud/koperator/pkg/resources/clusteripexternalaccess"
+	"github.com/banzaicloud/koperator/pkg/resources/contouringress"
 	"github.com/banzaicloud/koperator/pkg/resources/cruisecontrol"
 	"github.com/banzaicloud/koperator/pkg/resources/cruisecontrolmonitoring"
 	"github.com/banzaicloud/koperator/pkg/resources/envoy"
@@ -50,8 +50,9 @@ import (
 	"github.com/banzaicloud/koperator/pkg/resources/kafkamonitoring"
 	"github.com/banzaicloud/koperator/pkg/resources/nodeportexternalaccess"
 	"github.com/banzaicloud/koperator/pkg/util"
+
 	// TODO uncomment this
-	// contour "github.com/heptio/contour/apis/contour/v1beta1"
+	contour "github.com/projectcontour/contour/apis/projectcontour/v1"
 )
 
 var clusterFinalizer = "finalizer.kafkaclusters.kafka.banzaicloud.io"
@@ -119,7 +120,7 @@ func (r *KafkaClusterReconciler) Reconcile(ctx context.Context, request ctrl.Req
 		envoy.New(r.Client, instance),
 		istioingress.New(r.Client, instance),
 		nodeportexternalaccess.New(r.Client, instance),
-		clusteripexternalaccess.New(r.Client, instance),
+		contouringress.New(r.Client, instance),
 		kafkamonitoring.New(r.Client, instance),
 		cruisecontrolmonitoring.New(r.Client, instance),
 		kafka.New(r.Client, r.DirectClient, instance, r.KafkaClientProvider),
@@ -422,8 +423,8 @@ func envoyWatches(builder *ctrl.Builder) *ctrl.Builder {
 
 func contourWatches(builder *ctrl.Builder) *ctrl.Builder {
 	return builder.
-		Owns(&corev1.Service{})
-	// Owns(&contour.IngressRoute{})
+		Owns(&corev1.Service{}).
+		Owns(&contour.HTTPProxy{})
 }
 
 func cruiseControlWatches(builder *ctrl.Builder) *ctrl.Builder {
